Move teleop-vanilla -> vanilla
