+++ examples/vanilla/index.html
