+++ examples/vanilla/package.json
@@ -1,15 +1,17 @@
 {
   "name": "@viamrobotics/vanilla-sdk-example",
   "description": "A framework-free example application that demonstrates how to connect to a robot using @viamrobotics/sdk",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "license": "Apache-2.0",
   "scripts": {
     "preinstall": "make -C ../.. build",
-    "start": "esbuild ./src/main.ts --bundle --outfile=static/main.js --servedir=static",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "vite"
   },
-  "author": "",
-  "license": "Apache-2.0",
   "devDependencies": {
-    "esbuild": "0.16.12"
+    "typescript": "^5.2.2",
+    "vite": "^5.2.0"
   },
   "dependencies": {
     "@viamrobotics/sdk": "file:../.."
