+++ examples/vanilla/src/vite-env.d.ts
