--- examples/vanilla/static/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Viam Controls</title>
-    <link rel="icon" href="favicon.ico" />
-  </head>
-  <body>
-    <div id="main">
-      <button id="main-button" disabled=true>
-        Click me
-      </button>
-    </div>
-    <script type="module" src="main.js">
-    </script>
-  </body>
-</html>
