+++ examples/vanilla/tsconfig.json
